@@ -6,21 +6,21 @@ homepage: https://github.com/reduxsaga/counter
 repository: https://github.com/reduxsaga/counter
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  redux: ^4.0.0
-  flutter_redux: ^0.6.0
-  redux_saga: ^2.0.0
+  redux: ^5.0.0
+  flutter_redux: ^0.8.2
+  redux_saga: ^3.0.2
 
-  cupertino_icons: ^0.1.2
+  cupertino_icons: ^1.0.2
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  test: ^1.14.4
+  test: ^1.16.5
 
 flutter:
   uses-material-design: true
